pyproject.toml
[tool.poetry]
name = "diffbot-llm-inference"
version = "0.0.1"
description = "Diffbot LLM Inference Server"
authors = ["Diffbot <support@diffbot.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
openai = "^1.11.1"
python-dotenv = "^0.21.1"
pydantic = "^1.10.5"
tiktoken = "^0.5.2"
numpy = "^1.24.2"
pyyaml = "^6.0"
python-logstash = "^0.4.8"
httpx = "^0.24.1"
aiohttp = "^3.8.6"
setuptools = "^70.3.0"
unidecode = "^1.3.8"
fastapi-utils ="^0.7.0"
diskcache = "^5.6.3"
nodejs="^0.1.1"

[tool.poetry.scripts]
start = "server.main:start"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"
poetry = "^1.8.2"
poetry-plugin-export = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "tests",
]
python_files = [
  "test_*.py"
]
asyncio_mode="auto"
